Refactor tokenization and prefix handling
