+++ src/lib.rs
@@ -29,6 +29,7 @@ pub mod lists;
 pub mod process;
 mod reflow;
 pub mod table;
+mod tokenize;
 pub mod wrap;
 
 #[doc(hidden)]
